Rename messageReducers to messagesReducers.
